fix: Move option parsing into a separate module
