+++ src/parse-options.js
@@ -1,0 +1,26 @@
+import {resolve} from 'path'
+
+const DEFAULT_PROTRACTOR_ARGS = []
+const DEFAULT_OPTIONS = {
+  nodeBin: 'node',
+  maxAttempts: 3,
+  protractorArgs: DEFAULT_PROTRACTOR_ARGS,
+  parser: 'standard'
+}
+
+function parseOptions (providedOptions) {
+  let options = Object.assign({}, DEFAULT_OPTIONS, providedOptions)
+
+  if (options.protractorPath) {
+    options.protractorPath = resolve(options.protractorPath)
+  } else {
+    // '.../node_modules/protractor/lib/protractor.js'
+    let protractorMainPath = require.resolve('protractor')
+    // '.../node_modules/protractor/bin/protractor'
+    options.protractorPath = resolve(protractorMainPath, '../../bin/protractor')
+  }
+
+  return options
+}
+
+export default parseOptions
